chess_game_engine/requirements.txt
python-chess>=1.9.0
openai>=1.0.0
python-dotenv>=0.19.0


